Fix #119211: Masks do not update in GPU compositor
Masks are not updated when edited when using the GPU compositor. That's
because the GPU compositor caches static resources and invalidates them
according to the reclaculate flags that the depsgraph flushes to IDs.
The issue is that the flags are not flushed to the evaluated IDs of the
compositor depsgraph, but rather to some other evaluated versions of the
IDs.

To fix this, we also flush the recalculate flags to the draw data of
original IDs, and get them from original IDs in the compositor side. The
downside is that the draw data will not get freed until the ID is
deleted or the file is closed. But that is realisticly not an issue.