Fix: New sculpt brushes apply inverse deformation to deformed mesh
Currently we have to maintain an array of deformed positions on the
PBVH for historical reasons (once all brushes are refactored it can be
removed). The ported brushes handled it incorrectly though-- they
applied the inverse deformation to the translations applied to the
deformed mesh. Instead we have to apply the inverse deformation
after updating the PBVH's deformed positions.