Tests: Attempt at headless sculpt brush stroke performance test
As an initial step to creating automated regression tests for sculpt brushes,
try to make our existing performance test script into an automated test.
This doesn't work so far because the undo system isn't initialized in
background mode (and probably lots of other reasons). I'm not sure how
to resolve this yet.