Fix #74816: crash in Cycles light distribution due to integer overflow
`num_distribution` in `KernelIntegrator` has type `int`, which holds a
maximal value of 2147483647. However, when computing the distribution,
`size_t` is used, which can go beyond this value and result in a
negative value when converted to `int`.

This PR stops adding emissive triangles to the distribution when the
limit is reached. Those emitters can still be sampled during forward
path tracing, with a wrong MIS weight since they are not aware that
they cannot be sampled with NEE. We give a warning about the wrong
result upon calculating the distribution.

Also early return when `use_light_tree`. The block was there because
`num_distribution` was needed for light tree before bfd1836861.