Anim: add `ID*` cache of users to Action Bindings
Keep track of which IDs are animated by which Action Binding. This will be
necessary for display in the Action editor, where animation data that is
unrelated to the active object can be shown (when "show all bindings" is
on).

Note: animation evaluation will not be using this cache, at least not in
the near future. Potentially when we introduce animation-level constraints
this will change, but that's for the future.

--------------

This also adds an RNA function `action.bindings["name"].debug_log_users()`,
which can be used to print the list of users of that binding to stdout.
This is purely for debugging/testing this PR, and will be removed before
landing on `main`.

The user cache isn't actually used in this PR. It will be used soon in
#122672. That in itself is a big enough PR that I thought this aspect would
be nice to have as a separate one.