Color management: Support white balance as part of the display transform
This implements a von-Kries-style chromatic adaption using the Bradford matrix.
The adaption is performed in scene linear space in the OCIO GLSL shader, with
the matrix being computed on the host.

The parameters specify the white point of the input, which is to be mapped to
the white point of the scene linear space. The main parameter is temperature,
specified in Kelvin, which defines the blackbody spectrum that is used as the
input white point. Additionally, a tint parameter can be used to shift the
white point away from pure blackbody spectra (e.g. to match a D illuminant).

Two obvious future extensions are:
- Adding support for this to the compositor (should be quite easy, just call
  the same helper function and perform a matrix multiplication on each pixel).
- Adding support for picking temperature/tint from a given image. Performing
  the inverse calculation from white point to temperature/tint is doable, but
  implementing the color-picking itself in a way that accounts for e.g.
  the applied display transform is tricky.