Refactor: Anim filtering, avoid passing `bDopesheet *ads` parameter
In the animation filtering code, avoid passing `bDopesheet *ads` as a
separate parameter. Instead, pass `bAnimContext *ac`. This already
contains the same pointer. Some carefully placed asserts help to verify
that these pointers were indeed the same.

Not only do I think this approach is cleaner, it is a necessary step for
upcoming PRs #123187 and #122672. Those will need a `Main *bmain`
pointer, which is already available in the `bAnimContext` struct.

No functional changes.