Fix #104061: Python GPU module unusable after too early import
Add-ons may attempt to load the GPU module in backgroud mode when no GPU
context has been initialized yet. This would give an error on import.

If then later the GPU context does get initialized, for example for a
render engine, import would still fail as the module is cached.

This reverts commit d7f124f06f3fda59523ed05f12d1f4af6a523ba6, and again
throws errors in methods and constructors instead of module import.