VSE: further handle color tweaks
Selected handles always use the same color as the outline. Previously
they were almost always using it, except while actively transforming -
the outline was brighter but the handle itself was not.

With this, there's no need for handle colors state for the shader,
it is enough to put in flags which handles are selected vs which
ones are merely drawn.