Fix: VSE timeline strip readability issue due to lack of gaps
No actual report, but talked about a handful of times on the chat
recently. Especially with selected strips, it is not clear where
one of them begins and another ends since their outlines are
right next to each other. A solution was proposed to leave a one pixel
gap on the right side of each strip.