Fix: GPv3: Shortcut 'F' to change brush radius no longer working
This was introduced with b7c570c8543a246054e56a793096f6744662fdb3.
For the draw tools, we no longer use the unified paint settings.
This means that we can't use the `_template_paint_radial_control` anymore.

The fix explicitly sets up the shortcut to use the `brush.size` and `brush.strength`
for the `radial_control` operator.