Anim: prevent pushing down layered Actions onto the NLA
Prevent creating NLA strips from layered Actions via the 'Push Down'
operator.

Maybe in the future the NLA will support layered Actions (as a transitional
phase, before the NLA is replaced with layered Actions itself). For now,
it's better to create stable boundaries to prevent their use in the NLA.
When NLA support is being worked on, those boundaries can be removed again.