Python: Remove `color_attributes` property from unsupported IDs
Currently, the Python API for curves, pointclouds, and grease pencil IDs
has a `color_attributes` collection property (defined by `rna_def_attributes_common`).

Color attributes are special types of custom attributes that are only
supported for meshes. The code specifically checks for the mesh ID
type and does nothing otherwise.

This removes the API for the unsupported ID types. Note that this
technically breaks the API, but I can't think of a way this could have
been used in the first place.