Fix (unreported) MEM_new/MEM_freeN mismatch usages.
This one is a bit more involved than the previous ones, since the
mismatch is intentional here, and happens on a non-trivial type.

This happens because the new object (managed by the `unique_ptr`) steals
the internal data of the original object. Calling `MEM_delete` (and
therefore the destructor of the `AssetMetaData` object) would lead to
access-after-free and double-freeing errors.

Since this is a bit of a special case caused by the 'interface' between
traditional C-style of code in the FileBrowser, and actual Cpp code in
the asset area, a new dedicated function was added instead to the
`AssetMetaData` struct, which takes care of generating a `unique_ptr`
copy, and 'move' all internal data to it, leaving the original object
'empty', and safe to be deleted.

NOTE: we could also define a move assignement operator instead, however
this would hide the actual special case, think it's better to keep it
very visible and easy to find.