Fix #123716: autokeying Shape Key fields does not work
The issue had essentially the same origin as as #123538 (fixed
in #123719), but in the autokeying code. The autokeying code was
already fully resolving the RNA path to be keyed to be relative to the
owning ID, but the struct being keyed (even if it wasn't an ID) was
being passed to the keyframing code with that fully-resolved path.

This fixes it by simply passing the RNA pointer for the owning ID
rather than the struct being keyed.