Anim: give the `dependencies_id_types` for Actions a value dynamically
Instead of hard-coding `IDType_ID_AC.dependencies_id_types =
FILTER_ID_ALL`, which is overly broad, determine its value dynamically
in `id_type_init()` so that it's purely based on which `IDType`s
identify as 'animatable'.

This should make things like ID remapping a bit more efficient, as the
remapping code knows that an Action will only reference animatable
data-blocks.

Note that the initial value for the filter is still `FILTER_ID_ALL`, so
that in cases where (by accident) the call to `BKE_idtype_init()` is
omitted, semantically the situation is still correct. Not calling this
function is an easy to make oversight when writing unit tests.