macOS: Implement a Rosetta Splash Screen Warning
If Blender is running under Rosetta on macOS, meaning the user is using
an x86_64/Intel Blender binary on an arm64/Apple Silicon Mac, display
a splash screen warning to notify the user.

This is especially useful in the case where the user mistakenly
downloaded an Intel build of Blender, which can lead to reduced
performances due to emulation, with no other visible side effects.