VSE: Fix retiming menu polling
When the Strip -> Retiming menu was changed in #113211, the intended
behavior was for the menu items to be grayed out instead of hidden when
no retiming keys/strips were selected.

However, only "Add speed transition" was properly disabled, which meant
that the rest of the items remained visible whether or not items were
selected, and trying to use the operators in this state resulted in no
effect.

This patch properly grays out all retiming operators. The boolean
`is_retiming` in the strip retiming menu has been changed to
`is_retiming_mode`. This is intended to better indicate that all
retiming operators are polled away in their C++ code if nothing is
selected, but some are additionally polled away in the python UI code if
the user is in strip mode (editing strips) instead of retiming mode
(editing retiming keys).

This patch builds on #123973 which standardizes the functionality of the
reset retiming operator, meaning that the `not is_retiming_mode` poll is
no longer necessary. It also builds on #123975, which makes sure that
operators can run on active strips that disallow retiming (e.g. color
strip) as long as other strips that allow retiming are also selected.