Compositor: Implemenet Add Bloom Node Setup operator
This patch adds a new operator that adds a small compositor node setup
that replicates the old EEVEE Bloom effect. The setup should not be
expected to be identical to the old EEVEE bloom, because the bloom
algorithm itself changed, and some parameters were omitted.

Bloom was added to the Glare node as a replacement since Bloom was
removed in EEVEE Next. But users found it less intuitive and lost some
control like Tint and Intensity. So it was decided that a temporary
operator would be added to add a minimal node setup that tries to
replicate the old EEVEE bloom.

While the Intensity and Tint were exposed, the threshold parameters like
Clamp and Knee were omitted because they would otherwise be too complex
to create.