Fix: Titlebar header color on multi-editor floating windows
In the case of floating windows with multiples editors/areas, the window
titlebar would switch to the header color of whichever area was last
focused. Fixed by only applying the area header color to the titlebar
on single editor floating windows, and falling back to the header color
of the default space (SPACE_PROPERTIES) for multi-area windows otherwise.